+++ formulario/src/main/java/com/Aplicacion/App/Repository/ClienteRepository.java
@@ -1,10 +1,12 @@
 package com.Aplicacion.App.Repository;
 
+import java.util.ArrayList;
+
 import com.Aplicacion.App.Model.ClienteModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<ClienteModel, Integer> {
+public interface ClienteRepository extends JpaRepository<ClienteModel, Long> {
 
 }
